moved types outside of the serializers file
